+++ dtu_hpc_cli/history.py
@@ -66,6 +66,7 @@ class HistoryConfig:
     start_after: bool
     start_after_contains: str | None
     start_after_is: str | None
+    sync: bool
     walltime: bool
     walltime_above: Time | None
     walltime_below: Time | None
@@ -135,6 +136,8 @@ def execute_history(config: HistoryConfig):
         table.add_column("split_every")
     if config.start_after:
         table.add_column("start_after")
+    if config.sync:
+        table.add_column("sync")
     if config.branch:
         table.add_column("branch")
     if config.preamble:
@@ -176,6 +179,8 @@ def execute_history(config: HistoryConfig):
             row.append(str(values.split_every))
         if config.start_after:
             row.append(values.start_after if values.start_after is not None else "-")
+        if config.sync:
+            row.append("yes" if values.sync else "no")
         if config.branch:
             row.append(values.branch if values.branch is not None else "-")
         if config.preamble: