+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Run:
 
 Sync:
 * Delete files remotely if they have been removed locally.
+* Automatically sync when running install, submit, and resubmit (option to opt-out in config).
 
 ## v1.1.0